+++ dcspkg/src/commands/mod.rs
@@ -1,4 +1,5 @@
 mod install;
 mod list;
+mod run;
 
-pub use {install::install, list::list};
+pub use {install::install, list::list, run::run};
